content/canvas/test/webgl/conformance/gl-getstring.html

<!--
Copyright (c) 2009 The Chromium Authors. All rights reserved.
Use of this source code is governed by a BSD-style license that can be
found in the LICENSE file.
 -->
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
  "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
<title>WebGL gl.getString Conformance Tests</title>
<link rel="stylesheet" href="../resources/js-test-style.css"/>
<script src="../resources/desktop-gl-constants.js" type="text/javascript"></script>
<script src="../resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<canvas id="canvas" width="2" height="2"> </canvas>
<script>
description("This test checks getString returns strings in the correct format");

debug("");
debug("Canvas.getContext");

var gl = create3DContext(document.getElementById("canvas"));
if (!gl) {
  testFailed("context does not exist");
} else {
  testPassed("context exists");

  debug("");
  checkPrefix("OpenGL ES 2.0", "VERSION");
  checkPrefix("OpenGL ES GLSL ES 1.0", "SHADING_LANGUAGE_VERSION");
  shouldBe("gl.getString(gl.EXTENSIONS)", "''");
}

function checkPrefix(expected, enum) {
  var s = gl.getString(gl[enum]);
  if (s.length >= expected.length &&
      s.substring(0, expected.length) == expected) {
    testPassed("getString(gl." + enum + ") correctly started with " + expected);
  } else {
    testFailed("getString(gl." + enum + ") did not start with " + expected);
  }
}

debug("");
successfullyParsed = true;

</script>
<script src="../resources/js-test-post.js"></script>

<script>
</script>

</body>
</html>
